Moved modules and removed sys.path black magic
