+++ config.py.example
@@ -1,0 +1,39 @@
+##############################
+########## Discord ###########
+##############################
+
+ADMIN_ROLE = [] # Name of admin roles
+
+MUTED_ROLE = '' # Name of muted role
+
+GOTV_CHANNEL =  # ID of the channel on discord
+ANNOUNCEMENT =  # ID of the channel on discord
+
+GENERAL_LOGS =  # ID of the channel on discord
+COMMAND_LOGS =  # ID of the channel on discord
+
+TOKEN = '' # Token of discord BOT
+
+############################
+############eBot############
+############################
+
+URL_EBOT = '' # Base url of eBot
+
+EBOT_HOSTNAME = '' #  Hostname of eBot Database
+EBOT_PORT = '' # Port of eBot Database
+EBOT_DBNAME = '' # Dbname of eBot Database
+EBOT_USERNAME = ''  #Username of eBot Database
+EBOT_PASSWORD = '' #Password of eBot Database
+
+############################
+#########AdminAFK###########
+############################
+
+URL_ADMINAFK = '' #Base url of AdminAFK
+
+############################
+###########Other############
+############################
+
+RULES = '' #Rules link of the tournament
