+++ config/__init__.py
@@ -1,0 +1,2 @@
+from .config import *
+from .static_var import *
